Correct errors in the entity values to save in  azure table

+++ source/MyCloudProjectSample/MyCloudProject.Common/IExperimentResult.cs
@@ -16,7 +16,7 @@ namespace MyCloudProject.Common
         DateTime? StartTimeUtc { get; set; }
 
         DateTime? EndTimeUtc { get; set; }
-        float Accuracy { get; set; }
+        double Accuracy { get; set; }
     }
 
 }
